bug fixing

+++ x84/default/main.py
@@ -250,7 +250,7 @@ def renderscreen(menudraw=True, artdraw=True, bgdraw=True, tall=False, wide=Fals
         ypos += 1
 	#echo(background.pos(ypos, 1)+line)
     if len(linelist) < background.height - 1:
-       ypos = ((background.height - len(linelist)) / 2) + 1
+       ypos = int(((background.height - len(linelist)) / 2) + 1)
     else:
        ypos = 1
     for i in linelist:
